+++ libs/base_core/src/num/geom/dir/mod.rs
@@ -3,6 +3,8 @@
 #![doc = crate::_DOC_NUM_GEOM_DIR!()]
 //
 
+mod boundary; // Boundary1d, Boundary2d, Boundary3d
+// mod octant; // WIP
 mod orientation; // Orientation
 // mod radial_sectors;
 
@@ -10,6 +12,8 @@ crate::structural_mods! { // _mods
     _mods {
         #[doc(inline)]
         pub use super::{
+            boundary::*,
+            // octant::*, // WIP
             orientation::*,
             // radial_sectors::*;
         };