+++ libs/base_core/src/num/geom/mod.rs
@@ -6,6 +6,7 @@
 mod _helpers;
 
 pub mod dir; // Orientation
+// pub mod field; // WIP
 pub mod metric; // Distance, Extent, Orientation, Position...
 
 crate::structural_mods! { // _pub_mods
@@ -13,6 +14,7 @@ crate::structural_mods! { // _pub_mods
         #[doc(inline)]
         pub use super::{
             dir::_all::*,
+            // field::_all::*, // WIP
             metric::_all::*,
         };
     }